requirements.txt
streamlit>=1.32.0
langchain>=0.1.0
langchain-openai>=0.0.5
python-dotenv>=1.0.0
requests>=2.31.0
azure-openai>=1.0.0
